agent: Move unit tests about block device to block_device_handler
Move unit tests about block device to block_device_handler.
